+++ assistant.py
