--- setup.py
@@ -1,4 +1,0 @@
-from distutils.core import setup
-import py2exe
-
-setup(console=['main.py'])
